+++ src/Types.d.ts
@@ -81,6 +81,7 @@ export type TelnyxObject = {
   getInitialNetworkRetryDelay: () => number;
   getMaxNetworkRetryDelay: () => number;
   getMaxNetworkRetries: () => number;
+  getHttpsAgent: () => HttpsAgent;
   getConstant: <T = string>(name: string) => T;
   _setApiField: <K extends keyof TelnyxObject['_api']>(
     name: K,
@@ -180,7 +181,7 @@ export type TelnyxResourceObject = {
   _generateConnectionErrorMessage: (
     requestRetries: number,
   ) => string | undefined;
-  _shouldRetry: (res: ResponsePayload | null, numRetries: number) => boolean;
+  _shouldRetry: (res: ResponsePayload | null, numRetries: number, error?: TelnyxRawError) => boolean;
   _getSleepTimeInMS: (numRetries: number) => number;
   _defaultHeaders: (
     auth: RequestOptions['auth'],